Move the header of the controller to the upper layer
